+++ packages/plugin-image/web/src/createImagePlugin.js
@@ -48,6 +48,7 @@ const createImagePlugin = (config = {}) => {
       t,
       uiSettings,
       isMobile,
+      settings,
     }),
     helpers,
     anchorTarget,