+++ packages/plugin-image/web/src/toolbar/createToolbar.js
@@ -8,9 +8,17 @@ export default function createToolbar({
   relValue,
   uiSettings,
   isMobile,
+  settings,
 }) {
   return {
-    InlineButtons: createInlineButtons({ t, anchorTarget, relValue, uiSettings, isMobile }),
+    InlineButtons: createInlineButtons({
+      t,
+      anchorTarget,
+      relValue,
+      uiSettings,
+      isMobile,
+      imageEditorWixSettings: settings.imageEditorWixSettings,
+    }),
     InsertButtons: createInsertButtons({ helpers, t }),
     name: 'image',
   };