+++ packages/plugin-image/web/src/toolbar/inline-buttons.jsx
@@ -1,10 +1,29 @@
 import { BUTTONS, PluginSettingsIcon, getModalStyles } from 'wix-rich-content-common';
 import { Modals } from '../modals';
-import { MediaReplaceIcon } from '../icons';
+import { MediaReplaceIcon, ImageEditorIcon } from '../icons';
 
-export default ({ t, anchorTarget, relValue, uiSettings, isMobile }) => {
+const removeEmpty = list => list.filter(item => !!item);
+
+export default ({ t, anchorTarget, relValue, uiSettings, isMobile, imageEditorWixSettings }) => {
   const modalStyles = getModalStyles({ isMobile });
-  return [
+  const imageEditorStyles = getModalStyles({
+    customStyles: { content: { maxWidth: '100%', background: 'transparent' } },
+  });
+  const imageEditorButton = imageEditorWixSettings
+    ? {
+        keyName: 'imageEditor',
+        type: BUTTONS.EXTERNAL_MODAL,
+        icon: ImageEditorIcon,
+        modalName: Modals.IMAGE_EDITOR,
+        modalStyles: imageEditorStyles,
+        t,
+        imageEditorWixSettings,
+        mobile: false,
+        tooltipTextKey: 'ImageEditorButton_Tooltip',
+      }
+    : null;
+
+  const buttons = [
     { keyName: 'sizeOriginal', type: BUTTONS.SIZE_ORIGINAL, mobile: false },
     { keyName: 'sizeSmallCenter', type: BUTTONS.SIZE_SMALL_CENTER, mobile: false },
     { keyName: 'sizeContent', type: BUTTONS.SIZE_CONTENT, mobile: false },
@@ -15,6 +34,7 @@ export default ({ t, anchorTarget, relValue, uiSettings, isMobile }) => {
     { keyName: 'alignRight', type: BUTTONS.ALIGN_RIGHT, mobile: false },
     { keyName: 'separator2', type: BUTTONS.SEPARATOR, mobile: false },
     { keyName: 'link', type: BUTTONS.LINK, mobile: false },
+    imageEditorButton,
     {
       keyName: 'settings',
       type: BUTTONS.EXTERNAL_MODAL,
@@ -43,4 +63,6 @@ export default ({ t, anchorTarget, relValue, uiSettings, isMobile }) => {
     },
     { keyName: 'delete', type: BUTTONS.DELETE, mobile: true },
   ];
+
+  return removeEmpty(buttons);
 };